requirements.txt
discord.py
python-dotenv
langchain-community
langchain-openai
chromadb
PyPDF2
